requirements.txt
pyyaml==3.11
paramiko
argparse==1.3.0
git+https://github.com/mrts/sshtunnel.git@1a89e058abbeb5da199459e2c9103e419648bdfb
requests==2.8.1
mock==1.3.0
jinja2==2.8